requirements.txt
anyio==3.6.2
click==8.1.3
colorama==0.4.6
fastapi==0.92.0
greenlet==2.0.1
h11==0.14.0
pydantic==1.10.4
pyodbc==4.0.35
python-dotenv==0.21.1
SQLAlchemy==1.4.46
uvicorn==0.20.0


